fix: Update printf format specifiers for 64-bit integer ticks and Hz values

+++ main/core_portme.c
@@ -92,8 +92,8 @@ CORE_TICKS get_time(void) {
 
 secs_ret time_in_secs(CORE_TICKS ticks) {
     secs_ret retval = ((secs_ret)ticks) / ((secs_ret)EE_TICKS_PER_SEC);
-    ee_printf("Converting %" PRId64 " ticks at %" PRId64 " Hz: %.3f sec\n",
-              ticks, EE_TICKS_PER_SEC, (double)retval);
+    ee_printf("Converting %lld ticks at %lld Hz: %.3f sec\n",
+              (long long)ticks, (long long)EE_TICKS_PER_SEC, (double)retval);
     if (retval < 10.0) {
         ee_printf("WARNING: Runtime (%f secs) is less than required 10 seconds!\n", (double)retval);
     }
